Use highp and 32 bit as default

+++ Samples/Media/Hlms/PbsMobile/GLSL/PixelShader_ps.glsl
@@ -1,7 +1,7 @@
 @property( GL3+ )#version 330@end
 @property( !GL3+ )#define in varying@end
 /*#ifdef GL_ES
-precision mediump float;
+precision highp float;
 #endif*/
 #define FRAG_COLOR		0
 @property( !hlms_shadowcaster )
@@ -19,47 +19,47 @@ layout(location = FRAG_COLOR, index = 0) out vec4 outColour;
 @end
 
 @property( hlms_normal )
-in mediump vec3 psPos;
-in mediump vec3 psNormal;
-mediump vec3 nNormal;
-@property( normal_map )in mediump vec3 psTangent;
-mediump vec3 vTangent;@end
+in highp vec3 psPos;
+in highp vec3 psNormal;
+highp vec3 nNormal;
+@property( normal_map )in highp vec3 psTangent;
+highp vec3 vTangent;@end
 @end
 @property( hlms_pssm_splits )in highp float psDepth;@end
 @foreach( hlms_uv_count, n )
-in mediump vec@value( hlms_uv_count@n ) psUv@n;@end
+in highp vec@value( hlms_uv_count@n ) psUv@n;@end
 @foreach( hlms_num_shadow_maps, n )
 in highp vec4 psPosL@n;@end
 
 // START UNIFORM DECLARATION
 //Uniforms that change per pass
-@property( hlms_num_shadow_maps )uniform mediump vec2 invShadowMapSize[@value(hlms_num_shadow_maps)];
-uniform mediump float pssmSplitPoints[@value(hlms_pssm_splits)];@end
-@property( hlms_lights_spot )uniform mediump vec3 lightPosition[@value(hlms_lights_spot)];
-uniform mediump vec3 lightDiffuse[@value(hlms_lights_spot)];
-uniform mediump vec3 lightSpecular[@value(hlms_lights_spot)];
-@property(hlms_lights_attenuation)uniform mediump vec3 attenuation[@value(hlms_lights_attenuation)];@end
-@property(hlms_lights_spotparams)uniform mediump vec3 spotDirection[@value(hlms_lights_spotparams)];
-uniform mediump vec3 spotParams[@value(hlms_lights_spotparams)];@end
+@property( hlms_num_shadow_maps )uniform highp vec2 invShadowMapSize[@value(hlms_num_shadow_maps)];
+uniform highp float pssmSplitPoints[@value(hlms_pssm_splits)];@end
+@property( hlms_lights_spot )uniform highp vec3 lightPosition[@value(hlms_lights_spot)];
+uniform highp vec3 lightDiffuse[@value(hlms_lights_spot)];
+uniform highp vec3 lightSpecular[@value(hlms_lights_spot)];
+@property(hlms_lights_attenuation)uniform highp vec3 attenuation[@value(hlms_lights_attenuation)];@end
+@property(hlms_lights_spotparams)uniform highp vec3 spotDirection[@value(hlms_lights_spotparams)];
+uniform highp vec3 spotParams[@value(hlms_lights_spotparams)];@end
 @property( envprobe_map )uniform highp mat3 invViewMat; //This uniform depends on each renderable, but is the same for every pass@end
 @end
 
 //Uniforms that change per entity
-uniform lowp float roughness;
+uniform highp float roughness;
 /* kD is already divided by PI to make it energy conserving.
   (formula is finalDiffuse = NdotL * surfaceDiffuse / PI)
 */
-uniform lowp vec3 kD;
-uniform lowp vec3 kS;
-@property( fresnel_scalar )@piece( FresnelType )lowp vec3@end@end
-@property( !fresnel_scalar ) @piece( FresnelType )lowp float@end @end
+uniform highp vec3 kD;
+uniform highp vec3 kS;
+@property( fresnel_scalar )@piece( FresnelType )highp vec3@end@end
+@property( !fresnel_scalar ) @piece( FresnelType )highp float@end @end
 //Fresnel coefficient, may be per colour component (vec3) or scalar (float)
 uniform @insertpiece( FresnelType ) F0;
-@property( uv_atlas )uniform mediump vec3 atlasOffsets[@value( uv_atlas )];@end
-@property( normal_weight )uniform lowp float normalWeights[@value( normal_weight )];@end
-@property( detail_weights )uniform lowp vec4 cDetailWeights;@end
-@property( detail_offsetsD )uniform mediump vec4 detailOffsetScaleD[@value( detail_offsetsD )];@end
-@property( detail_offsetsN )uniform mediump vec4 detailOffsetScaleN[@value( detail_offsetsN )];@end
+@property( uv_atlas )uniform highp vec3 atlasOffsets[@value( uv_atlas )];@end
+@property( normal_weight )uniform highp float normalWeights[@value( normal_weight )];@end
+@property( detail_weights )uniform highp vec4 cDetailWeights;@end
+@property( detail_offsetsD )uniform highp vec4 detailOffsetScaleD[@value( detail_offsetsD )];@end
+@property( detail_offsetsN )uniform highp vec4 detailOffsetScaleN[@value( detail_offsetsN )];@end
 // END UNIFORM DECLARATION
 
 @property( detail_offsetsD )
@@ -78,27 +78,27 @@ uniform @insertpiece( FresnelType ) F0;
 @end
 
 @property( !roughness_map )#define ROUGHNESS roughness@end
-@property( diffuse_map )uniform lowp sampler2D	texDiffuseMap;@end
-@property( normal_map_tex )uniform lowp sampler2D	texNormalMap;@end
-@property( specular_map )uniform lowp sampler2D	texSpecularMap;@end
-@property( roughness_map )uniform lowp sampler2D	texRoughnessMap;@end
+@property( diffuse_map )uniform highp sampler2D	texDiffuseMap;@end
+@property( normal_map_tex )uniform highp sampler2D	texNormalMap;@end
+@property( specular_map )uniform highp sampler2D	texSpecularMap;@end
+@property( roughness_map )uniform highp sampler2D	texRoughnessMap;@end
 @property( envprobe_map )
-@property( !hlms_cube_arrays_supported )uniform lowp samplerCube	texEnvProbeMap;@end
-@property( !hlms_cube_arrays_supported && false )uniform lowp sampler2D	texEnvProbeMap;@end @end
-@property( detail_weight_map )uniform lowp sampler2D	texDetailWeightMap;@end
-@property( detail_maps_diffuse )uniform lowp sampler2D	texDetailMap[@value( detail_maps_diffuse )];@end
-@property( detail_maps_normal )uniform lowp sampler2D	texDetailNormalMap[@value( detail_maps_normal )];@end
+@property( !hlms_cube_arrays_supported )uniform highp samplerCube	texEnvProbeMap;@end
+@property( !hlms_cube_arrays_supported && false )uniform highp sampler2D	texEnvProbeMap;@end @end
+@property( detail_weight_map )uniform highp sampler2D	texDetailWeightMap;@end
+@property( detail_maps_diffuse )uniform highp sampler2D	texDetailMap[@value( detail_maps_diffuse )];@end
+@property( detail_maps_normal )uniform highp sampler2D	texDetailNormalMap[@value( detail_maps_normal )];@end
 
-@property( !diffuse_map && detail_maps_diffuse )lowp vec4 diffuseCol;@end
-@property( diffuse_map )lowp vec4 diffuseCol;
+@property( !diffuse_map && detail_maps_diffuse )highp vec4 diffuseCol;@end
+@property( diffuse_map )highp vec4 diffuseCol;
 @piece( SampleDiffuseMap )	diffuseCol = @insertpiece(texture2D)( texDiffuseMap, psUv@value(uv_diffuse).xy * atlasOffsets[@value(atlas)].z + atlasOffsets[@counter(atlas)].xy );
 @property( !hw_gamma_read )	//Gamma to linear space
 	diffuseCol = diffuseCol * diffuseCol;@end @end
 @piece( MulDiffuseMapValue )* diffuseCol.xyz@end@end
-@property( specular_map )lowp vec3 specularCol;
+@property( specular_map )highp vec3 specularCol;
 @piece( SampleSpecularMap )	specularCol = @insertpiece(texture2D)( texSpecularMap, psUv@value(uv_specular).xy * atlasOffsets[@value(atlas)].z + atlasOffsets[@counter(atlas)].xy ).xyz;@end
 @piece( MulSpecularMapValue )* specularCol@end@end
-@property( roughness_map )lowp float ROUGHNESS;
+@property( roughness_map )highp float ROUGHNESS;
 @piece( SampleRoughnessMap )ROUGHNESS = roughness * @insertpiece(texture2D)( texRoughnessMap, psUv@value(uv_roughness).xy * atlasOffsets[@value(atlas)].z + atlasOffsets[@counter(atlas)].xy ).x;@end
 @end
 
@@ -107,11 +107,11 @@ uniform @insertpiece( FresnelType ) F0;
 @property( !hlms_shadow_uses_depth_texture )@piece( SAMPLER2DSHADOW )sampler2D@end @end
 uniform highp @insertpiece( SAMPLER2DSHADOW ) texShadowMap[@value(hlms_num_shadow_maps)];
 
-lowp float getShadow( highp @insertpiece( SAMPLER2DSHADOW ) shadowMap, highp vec4 psPosLN, mediump vec2 invShadowMapSize )
+highp float getShadow( highp @insertpiece( SAMPLER2DSHADOW ) shadowMap, highp vec4 psPosLN, highp vec2 invShadowMapSize )
 {
 @property( !hlms_shadow_usues_depth_texture )
 	highp float fDepth = psPosLN.z;
-	mediump vec2 uv = psPosLN.xy / psPosLN.w;
+	highp vec2 uv = psPosLN.xy / psPosLN.w;
 	highp vec3 o = vec3( invShadowMapSize, -invShadowMapSize.x ) * 0.3;
 
 	// 2x2 PCF
@@ -126,7 +126,7 @@ lowp float getShadow( highp @insertpiece( SAMPLER2DSHADOW ) shadowMap, highp vec
 }
 @end
 
-@property( hlms_lights_spot_textured )mediump vec3 zAxis( mediump vec4 qQuat )
+@property( hlms_lights_spot_textured )highp vec3 zAxis( highp vec4 qQuat )
 {
 	Real fTy  = 2.0 * qQuat.y;
 	Real fTz  = 2.0 * qQuat.z;
@@ -137,17 +137,17 @@ lowp float getShadow( highp @insertpiece( SAMPLER2DSHADOW ) shadowMap, highp vec
 	Real fTyy = fTy * qQuat.y;
 	Real fTzz = fTz * qQuat.z;
 
-	return mediump vec3( 1.0-(fTyy+fTzz), fTxy+fTwz, fTxz-fTwy );
+	return highp vec3( 1.0-(fTyy+fTzz), fTxy+fTwz, fTxz-fTwy );
 }
-mediump vec3 qmul( mediump vec4 q, mediump vec3 v )
+highp vec3 qmul( highp vec4 q, highp vec3 v )
 {
 	return v + 2.0 * cross( cross( v, q.xyz ) + q.w * v, q.xyz );
 }
 @end
 
-@property( normal_map_tex )mediump vec3 getTSNormal( lowp sampler2D normalMap, lowp vec2 uv )
+@property( normal_map_tex )highp vec3 getTSNormal( highp sampler2D normalMap, highp vec2 uv )
 {
-	mediump vec3 tsNormal;
+	highp vec3 tsNormal;
 @property( signed_int_textures )
 	//Normal texture must be in U8V8 or BC5 format!
 	tsNormal.xy = @insertpiece(texture2D)( normalMap, uv * atlasOffsets[@value(atlas)].z + atlasOffsets[@counter(atlas)].xy ).xy;
@@ -160,9 +160,9 @@ mediump vec3 qmul( mediump vec4 q, mediump vec3 v )
 	return tsNormal;
 }
 @end
-@property( detail_maps_normal )mediump vec3 getTSDetailNormal( lowp sampler2D normalMap, lowp vec2 uv )
+@property( detail_maps_normal )highp vec3 getTSDetailNormal( highp sampler2D normalMap, highp vec2 uv )
 {
-	mediump vec3 tsNormal;
+	highp vec3 tsNormal;
 @property( signed_int_textures )
 	//Normal texture must be in U8V8 or BC5 format!
 	tsNormal.xy = @insertpiece(texture2D)( normalMap, uv ).xy;
@@ -187,21 +187,21 @@ mediump vec3 qmul( mediump vec4 q, mediump vec3 v )
 	@end
 @end
 
-mediump vec3 cookTorrance( mediump vec3 lightDir, mediump vec3 viewDir, lowp float NdotV, mediump vec3 lightDiffuse, mediump vec3 lightSpecular )
+highp vec3 cookTorrance( highp vec3 lightDir, highp vec3 viewDir, highp float NdotV, highp vec3 lightDiffuse, highp vec3 lightSpecular )
 {
-	mediump vec3 halfWay= normalize( lightDir + viewDir );
-	lowp float NdotL = clamp( dot( nNormal, lightDir ), 0.0, 1.0 );
-	lowp float NdotH = clamp( dot( nNormal, halfWay ), 0.001, 1.0 );
-	lowp float VdotH = clamp( dot( viewDir, halfWay ), 0.001, 1.0 );
+	highp vec3 halfWay= normalize( lightDir + viewDir );
+	highp float NdotL = clamp( dot( nNormal, lightDir ), 0.0, 1.0 );
+	highp float NdotH = clamp( dot( nNormal, halfWay ), 0.001, 1.0 );
+	highp float VdotH = clamp( dot( viewDir, halfWay ), 0.001, 1.0 );
 
-	lowp float sqR = ROUGHNESS * ROUGHNESS;
+	highp float sqR = ROUGHNESS * ROUGHNESS;
 
 	//Roughness term (Beckmann distribution)
 	//Formula:
 	//	Where alpha = NdotH and m = roughness
 	//	R = [ 1 / (m^2 x cos(alpha)^4 ] x [ e^( -tan(alpha)^2 / m^2 ) ]
 	//	R = [ 1 / (m^2 x cos(alpha)^4 ] x [ e^( ( cos(alpha)^2 - 1 )  /  (m^2 cos(alpha)^2 ) ]
-	mediump float NdotH_sq = NdotH * NdotH;
+	highp float NdotH_sq = NdotH * NdotH;
 	highp float roughness_a = 1.0 / ( 3.141592654 * sqR * NdotH_sq * NdotH_sq );//( 1 / (m^2 x cos(alpha)^4 )
 	highp float roughness_b = NdotH_sq - 1.0;	//( cos(alpha)^2 - 1 )
 	highp float roughness_c = sqR * NdotH_sq;		//( m^2 cos(alpha)^2 )
@@ -235,42 +235,42 @@ void main()
 {
 @property( detail_maps_diffuse || detail_maps_normal )
 	@property( detail_weight_map )
-		lowp vec4 detailWeights = @insertpiece(texture2D)( texDetailWeightMap, psUv@value(uv_detail_weight).xy );
+		highp vec4 detailWeights = @insertpiece(texture2D)( texDetailWeightMap, psUv@value(uv_detail_weight).xy );
 		@property( detail_weights )detailWeights *= cDetailWeights;@end
 	@end @property( !detail_weight_map )
-		@property( detail_weights )lowp vec4 detailWeights = cDetailWeights;@end
-		@property( !detail_weights )lowp vec4 detailWeights = vec4( 1.0 );@end
+		@property( detail_weights )highp vec4 detailWeights = cDetailWeights;@end
+		@property( !detail_weights )highp vec4 detailWeights = vec4( 1.0 );@end
 	@end
 	//Group all texture loads together to help the GPU hide the
 	//latency (bad GL ES2 drivers won't optimize this automatically)
 @end
 
 @foreach( detail_maps_diffuse, n )
-	lowp vec4 detailCol@n	= @insertpiece(texture2D)( texDetailMap[@n], psUv@value(uv_detail@n).xy@insertpiece( offsetDetailD@n ) );
+	highp vec4 detailCol@n	= @insertpiece(texture2D)( texDetailMap[@n], psUv@value(uv_detail@n).xy@insertpiece( offsetDetailD@n ) );
 	@property( !hw_gamma_read )//Gamma to linear space
 		detailCol.xyz = detailCol.xyz * detailCol.xyz;@end
 	detailWeights.@insertpiece(detail_diffuse_swizzle@n) *= detailCol@n.w;
 	detailCol@n.w = detailWeights.@insertpiece(detail_diffuse_swizzle@n);
 @end
 @foreach( detail_maps_normal, n )
-	mediump vec3 vDetail@n	= getTSDetailNormal( texDetailNormalMap[@n], psUv@value(uv_detail_nm@n).xy@insertpiece( offsetDetailN@n ) ) * detailWeights.@insertpiece(detail_normal_swizzle@n) @insertpiece( detail@n_nm_weight_mulA );@end
+	highp vec3 vDetail@n	= getTSDetailNormal( texDetailNormalMap[@n], psUv@value(uv_detail_nm@n).xy@insertpiece( offsetDetailN@n ) ) * detailWeights.@insertpiece(detail_normal_swizzle@n) @insertpiece( detail@n_nm_weight_mulA );@end
 
 @property( !normal_map )
 	nNormal = normalize( psNormal );
 @end @property( normal_map )
-	mediump vec3 geomNormal = normalize( psNormal );
+	highp vec3 geomNormal = normalize( psNormal );
 	vTangent = normalize( psTangent );
 
 	//Get the TBN matrix
-	mediump vec3 vBinormal	= cross( vTangent, geomNormal );
-	mediump mat3 TBN		= mat3( vTangent, vBinormal, geomNormal );
+	highp vec3 vBinormal	= cross( vTangent, geomNormal );
+	highp mat3 TBN		= mat3( vTangent, vBinormal, geomNormal );
 
 	@property( normal_map_tex )nNormal = getTSNormal( texNormalMap, psUv@value(uv_normal).xy );@end
 	@property( normal_weight_tex )nNormal = mix( vec3( 0.0, 0.0, 1.0 ), nNormal, normalWeights[0] );@end
 @end
 
 @property( hlms_pssm_splits )
-	lowp float fShadow = 1.0;
+	highp float fShadow = 1.0;
 	if( psDepth <= pssmSplitPoints[@value(CurrentShadowMap)] )
 		fShadow = getShadow( texShadowMap[@value(CurrentShadowMap)], psPosL0, invShadowMapSize[@counter(CurrentShadowMap)] );@end
 @foreach( hlms_pssm_splits, n, 1 )	else if( psDepth <= pssmSplitPoints[@value(CurrentShadowMap)] )
@@ -307,10 +307,10 @@ void main()
 
 	//Everything's in Camera space, we use Cook-Torrance lighting
 @property( hlms_lights_spot || envprobe_map )
-	mediump vec3 viewDir	= normalize( -psPos );
-	lowp float NdotV		= clamp( dot( nNormal, viewDir ), 0.0, 1.0 );@end
+	highp vec3 viewDir	= normalize( -psPos );
+	highp float NdotV		= clamp( dot( nNormal, viewDir ), 0.0, 1.0 );@end
 
-	mediump vec3 finalColour = vec3(0);
+	highp vec3 finalColour = vec3(0);
 @property( hlms_lights_directional )
 	finalColour += cookTorrance( lightPosition[0], viewDir, NdotV, lightDiffuse[0], lightSpecular[0] );
 @property( hlms_num_shadow_maps )	finalColour *= fShadow;	//1st directional light's shadow@end
@@ -318,10 +318,10 @@ void main()
 @foreach( hlms_lights_directional, n, 1 )
 	finalColour += cookTorrance( lightPosition[@n], viewDir, NdotV, lightDiffuse[@n], lightSpecular[@n] )@insertpiece( DarkenWithShadow );@end
 
-@property( hlms_lights_point || hlms_lights_spot )	mediump vec3 lightDir;
-	mediump float fDistance;
-	mediump vec3 tmpColour;
-	mediump float spotCosAngle;@end
+@property( hlms_lights_point || hlms_lights_spot )	highp vec3 lightDir;
+	highp float fDistance;
+	highp vec3 tmpColour;
+	highp float spotCosAngle;@end
 
 	//Point lights
 @foreach( hlms_lights_point, n, hlms_lights_directional )
@@ -331,7 +331,7 @@ void main()
 	{
 		lightDir *= 1.0 / fDistance;
 		tmpColour = cookTorrance( lightDir, viewDir, NdotV, lightDiffuse[@n], lightSpecular[@n] )@insertpiece( DarkenWithShadow );
-		mediump float atten = 1.0 / (1.0 + attenuation[@value(atten)].y * fDistance + attenuation[@counter(atten)].z * fDistance * fDistance );
+		highp float atten = 1.0 / (1.0 + attenuation[@value(atten)].y * fDistance + attenuation[@counter(atten)].z * fDistance * fDistance );
 		finalColour += tmpColour * atten;
 	}@end
 
@@ -348,21 +348,21 @@ void main()
 	{
 		lightDir *= 1.0 / fDistance;
 	@property( hlms_lights_spot_textured )
-		mediump vec3 posInLightSpace = qmul( spotQuaternion[@value(spot_params)], psPos );
-		lowp float spotAtten = @insertpiece(texture2D)( texSpotLight, normalize( posInLightSpace ).xy ).x;
+		highp vec3 posInLightSpace = qmul( spotQuaternion[@value(spot_params)], psPos );
+		highp float spotAtten = @insertpiece(texture2D)( texSpotLight, normalize( posInLightSpace ).xy ).x;
 	@end
 	@property( !hlms_lights_spot_textured )
-		mediump float spotAtten = clamp( (spotCosAngle - spotParams[@value(spot_params)].y) * spotParams[@value(spot_params)].x, 0.0, 1.0 );
+		highp float spotAtten = clamp( (spotCosAngle - spotParams[@value(spot_params)].y) * spotParams[@value(spot_params)].x, 0.0, 1.0 );
 		spotAtten = pow( spotAtten, spotParams[@counter(spot_params)].z );
 	@end
 		tmpColour = cookTorrance( lightDir, viewDir, NdotV, lightDiffuse[@n], lightSpecular[@n] )@insertpiece( DarkenWithShadow );
-		mediump float atten = 1.0 / (1.0 + attenuation[@value(atten)].y * fDistance + attenuation[@counter(atten)].z * fDistance * fDistance );
+		highp float atten = 1.0 / (1.0 + attenuation[@value(atten)].y * fDistance + attenuation[@counter(atten)].z * fDistance * fDistance );
 		finalColour += tmpColour * (atten * spotAtten);
 	}@end
 
 @property( envprobe_map )
-	mediump vec3 reflDir = 2.0 * dot( viewDir, nNormal ) * nNormal - viewDir;
-	mediump vec3 envColour = @insertpiece(textureCubeLod)( texEnvProbeMap, invViewMat * reflDir, ROUGHNESS * 12.0 ).xyz;
+	highp vec3 reflDir = 2.0 * dot( viewDir, nNormal ) * nNormal - viewDir;
+	highp vec3 envColour = @insertpiece(textureCubeLod)( texEnvProbeMap, invViewMat * reflDir, ROUGHNESS * 12.0 ).xyz;
 	@property( !hw_gamma_read )//Gamma to linear space
 	envColour = envColour * envColour;@end
 	finalColour += cookTorrance( reflDir, viewDir, NdotV, envColour, envColour * (ROUGHNESS * ROUGHNESS) );@end
